+++ src/test/java/io/electrum/vas/api/ResourceTestInterfaceImpl.java
@@ -1,0 +1,12 @@
+package io.electrum.vas.api;
+
+public class ResourceTestInterfaceImpl implements ResourceTestInterface {
+
+   @Override
+   public void correctBaseMethod(Object obj) {
+   }
+
+   @Override
+   public void incorrectBaseMethod(Object obj) {
+   }
+}